Simple 64bit hotfix
git-svn-id: http://svn.clifford.at/libxsvf/trunk@66 26ca6efd-f666-0410-bd9f-983cc4be47ed

+++ xsvftool-ft2232h.c
@@ -708,7 +708,7 @@ static void help()
 	fprintf(stderr, "Lib(X)SVF is free software licensed under the ISC license.\n");
 	fprintf(stderr, "\n");
 	fprintf(stderr, "Usage: %s [ -v[v..] ] [ -d dumpfile ] [ -L | -B ] [ -f freq[k|M] ] \\\n", progname);
-	fprintf(stderr, "      %*s { -s svf-file | -x xsvf-file | -c } ...\n", strlen(progname)+1, "");
+	fprintf(stderr, "      %*s { -s svf-file | -x xsvf-file | -c } ...\n", (int)(strlen(progname)+1), "");
 	fprintf(stderr, "\n");
 	fprintf(stderr, "   -v\n");
 	fprintf(stderr, "          Enable verbose output (repeat for incrased verbosity)\n");
